Patch introduces the changes to run performance tests in RHTAP.
Within RHTAP, an ephemeral environment (namespace), will be used to temporarily deploy and run performance tests.A custom deployment, a devfile.yaml that points to  Kubernetes YAML(deploy.yaml) file and Dockerfile(Containerfile.rhtap), will be used for deployment to the ephemeral environment. The pipeline .tekton/rhtap-perf-integration-test.yaml runs performance test in the ephemeral environment and does regression analysis by querying and  pushing the results to the horreum.