Add clone validation e2e test
Test includes the first and sixth steps as defined in the spec. The test is using the clone script to make a clone and validating that expected directories exist. It also implements the cleanup step in a way that should be scalable to the future planned additions to the test.

Note that the test is currently failing due to a bug in the clone script where it expects to be run from within the scripts/clone/ directory despite documentation stating to run it from the project root.